Move SiblingSubgraph tests into new file
